+++ packages/plugin-audio/web/src/editor.ts
@@ -1,13 +1,8 @@
-/*
-  This module exports the required CreatePluginFunction for RicosEditor.
-  If your plugin uses a modal, then make sure to uncomment 'ModalsMap'.
-*/
+import { createAudioData } from './createAudioData';
 
 import { createAudioPlugin } from './createAudioPlugin';
 import type { AudioPluginEditorConfig } from './types';
 import { AUDIO_TYPE } from './types';
-// import { ModalsMap } from './modals'; // Optional
-// import { theme } from './defaults'; // Optional
 import { DEFAULTS } from './defaults';
 import type { EditorPluginCreator } from 'wix-rich-content-common';
 
@@ -16,7 +11,6 @@ export const pluginAudio: EditorPluginCreator<AudioPluginEditorConfig> = config 
     config: { ...DEFAULTS.config, ...config },
     type: AUDIO_TYPE,
     createPlugin: createAudioPlugin,
-    // ModalsMap,
-    // theme,
+    createPluginData: createAudioData,
   };
 };