+++ packages/plugin-audio/web/src/types.ts
@@ -1,6 +1,5 @@
 import type { EditorPluginConfig, ViewerPluginConfig } from 'wix-rich-content-common';
-
-export const AUDIO_TYPE = 'audio';
+export { AUDIO_TYPE } from 'ricos-content';
 
 export interface AudioPluginEditorConfig extends EditorPluginConfig {
   [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any