+++ packages/ricos-content/web/src/consts.ts
@@ -73,6 +73,7 @@ export const UNDO_REDO_TYPE = 'wix-rich-content-undo-redo';
 export const APP_EMBED_TYPE = 'wix-draft-plugin-vertical-embed';
 export const VERTICAL_EMBED_TYPE = 'wix-draft-plugin-vertical-embed';
 export const VIDEO_TYPE = 'wix-draft-plugin-video';
+export const AUDIO_TYPE = 'ricos-audio';
 export const VIDEO_TYPE_LEGACY = 'VIDEO-EMBED';
 export const POLL_TYPE = 'wix-draft-plugin-poll';
 export const TABLE_TYPE = 'wix-rich-content-plugin-table';
@@ -128,6 +129,7 @@ export const TIPTAP_GALLERY_TYPE = 'gallery';
 export const TIPTAP_FILE_TYPE = 'file';
 export const TIPTAP_GIF_TYPE = 'gif';
 export const TIPTAP_VIDEO_TYPE = 'video';
+export const TIPTAP_AUDIO_TYPE = 'audio';
 export const TIPTAP_LINK_PREVIEW_TYPE = 'linkPreview';
 export const TIPTAP_POLL_TYPE = 'poll';
 export const TIPTAP_APP_EMBED_TYPE = 'appEmbed';
@@ -155,4 +157,5 @@ export const RICOS_NODE_TYPE_TO_DATA_FIELD = {
   [Node_Type.BLOCKQUOTE]: 'paragraphData',
   [Node_Type.EMBED]: 'embedData',
   [Node_Type.EXTERNAL]: 'externalData',
+  [Node_Type.AUDIO]: 'audioData',
 } as const;