+++ packages/ricos-content/web/src/converters/plainText/toPlainText/complexPlainText.ts
@@ -81,4 +81,7 @@ Yo yo yo!
 
 
 
+
+Dear Fear
+https://static.wixstatic.com/mp3/f0f74f_35a1cdce4973490eac49e74c3244364d.mp3
 `;