+++ packages/ricos-content/web/src/types/contentApi.ts
@@ -17,6 +17,7 @@ import type {
   TableData,
   TextData,
   VideoData,
+  AudioData,
   TableCellData,
   CollapsibleListData,
   PollData,
@@ -99,6 +100,7 @@ type AddMap = {
   addHtml: HTMLData;
   addImage: ImageData;
   addVideo: VideoData;
+  addAudio: AudioData;
   addAppEmbed: AppEmbedData;
   addEmbed: EmbedData;
   addLinkPreview: LinkPreviewData;
@@ -195,6 +197,7 @@ type RenamedBuilderFunctions = PickRenameMulti<
     addPoll: 'poll';
     addTable: 'table';
     addVideo: 'video';
+    addAudio: 'audio';
   }
 >;